Fully rewrite login and DIS functions to avoid generating multiple and keeping expired SIDs.
Addresses #34:
Previous versions generated at least three separate SIDs for the
discovery, the bridge and the threads. These SIDs were never logged out
properly. They could be reused upon Surveillance Station restart, which
_was_ tested before. Upon Disk Station restart all SIDs became expired /
invalid and there was no way to renew them apart of restarting the
binding. That affected specially the HomeMode thread, which had -
unintended - its own SID.

Note: for reusing SID, logout function had to be disabled as of now.
It's no problem as it had never worked as intended before...

As it's a rather critical change, heavy testing is required.